Move transaction into domain module
